Split device enumeration into a separate feature
Prior to this, enumerating USB bite healers and establishing a
connection to them happened in the same generator step.

Split those concerns so that enumerating happens first, producing a
metadata object as an intermediate result.

The metadata object sets the stage for the `info` subcommand.
It also provides a `connect()` method, which establishes the actual
connection.